+++ samples/sample-express/src/models/UpdateUserDto.ts
@@ -1,5 +1,7 @@
 import {IsNotEmpty, IsString, MaxLength, MinLength} from "class-validator";
+import {Model} from "@nodeboot/starter-openapi";
 
+@Model()
 export class UpdateUserDto {
     @IsString()
     @IsNotEmpty()