+++ samples/sample-fastify/src/auth/LoggedInUserResolver.ts
@@ -1,7 +1,6 @@
 import {Action, CurrentUserChecker} from "@nodeboot/context";
 import {Component} from "@nodeboot/core";
-import {FastifyRequest} from "fastify/types/request";
-import {FastifyReply} from "fastify/types/reply";
+import {FastifyReply, FastifyRequest} from "fastify";
 import {Inject} from "@nodeboot/di";
 import {Logger} from "winston";