+++ packages/context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/context
 
+## 1.12.0
+
+### Minor Changes
+
+-   Upgrade typescript, fix eslint and fix aot openapi model generation
+
 ## 1.11.1
 
 ### Patch Changes