+++ packages/di/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/di
 
+## 1.4.0
+
+### Minor Changes
+
+-   Upgrade typescript, fix eslint and fix aot openapi model generation
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.12.0
+
 ## 1.3.1
 
 ### Patch Changes