+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/tools
 
+## 1.2.0
+
+### Minor Changes
+
+-   Upgrade typescript, fix eslint and fix aot openapi model generation
+
 ## 1.1.0
 
 ### Minor Changes