+++ samples/sample-ghost-server/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @nodeboot/express-ghost-server
 
+## 2.3.1
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/aot@1.4.0
+    -   @nodeboot/config@1.3.0
+    -   @nodeboot/context@1.12.0
+    -   @nodeboot/core@1.14.0
+    -   @nodeboot/di@1.4.0
+    -   @nodeboot/error@1.3.0
+    -   @nodeboot/ghost-server@1.2.0
+    -   @nodeboot/starter-http@3.5.0
+    -   @nodeboot/starter-persistence@1.10.0
+    -   @nodeboot/starter-scheduler@1.5.0
+
 ## 2.3.0
 
 ### Minor Changes