+++ servers/fastify-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/fastify-server
 
+## 1.7.1
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.12.0
+    -   @nodeboot/core@1.14.0
+    -   @nodeboot/engine@1.5.0
+    -   @nodeboot/error@1.3.0
+
 ## 1.7.0
 
 ### Minor Changes