+++ servers/ghost-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/ghost-server
 
+## 1.2.0
+
+### Minor Changes
+
+-   Upgrade typescript, fix eslint and fix aot openapi model generation
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.12.0
+    -   @nodeboot/core@1.14.0
+    -   @nodeboot/engine@1.5.0
+    -   @nodeboot/error@1.3.0
+
 ## 1.1.6
 
 ### Patch Changes