+++ starters/scheduler/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/starter-scheduler
 
+## 1.5.0
+
+### Minor Changes
+
+-   Upgrade typescript, fix eslint and fix aot openapi model generation
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.12.0
+
 ## 1.4.1
 
 ### Patch Changes